+++ Astromatch/projetoastromatch/src/App.js
@@ -1,32 +1,16 @@
 import React, { useEffect, useState} from 'react'
-import Main from './Main'
+import Main from './components/Main'
 import PageMatches from './components/PageMatches'
 
 function App() {
 
   const [page, setPage] = useState(false)
-  // const [view, setView] = useState (true)
-  
-  // const changePage = ()=>{
-  //   setView (false)
-  // }
-
-  // const definePage = ()=>{
-  // {view?<Main/>:<PageMatches/>}
-  // }
-
-  // useEffect (() =>{    
-  //   definePage()
-  // })
-
   const stayMain = () => {
     setPage (true)
   }
-
   const changeToMatches = () => {
     setPage (false)
   }
-
   const setScreen =() =>{
     if (page){
       return (
@@ -37,18 +21,12 @@ function App() {
         <Main stayMain = {stayMain}/>
 )}}
 
-
   useEffect (() =>{    
     setScreen()
   })
-
   return (
   <div>
-   
-    {/* <button onClick={definePage}>Duas Páginas</button>     */}
-
     <div>{setScreen()}</div>
-        
   </div>
   );
 }
